backend/requirements.txt
flask
flask-bcrypt
flask-login
mysql-connector-python
python-dotenv
